+++ src/test/java/scootertest/chrometests/ConnectionChrome.java
@@ -1,5 +1,6 @@
-package scooterTest.chromeTests;
+package scootertest.chrometests;
 
+import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.After;
 import org.junit.Before;
 import org.openqa.selenium.WebDriver;
@@ -12,9 +13,11 @@ public class ConnectionChrome {
     WebDriver driver;
 
     //подключение драйвера браузера и установление ожидания в 5 секунд
+
     @Before
-    public void open() {
-        System.setProperty("webdriver.chrome.driver","C:\\WebDriver\\bin\\chromedriver.exe");
+    public void setupBrowser() {
+        // Автоматическое управление драйвером Chrome
+        WebDriverManager.chromedriver().setup();
         driver = new ChromeDriver();
         driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
     }
